Completed side-task 2 for Issue #23. Implemented the Wage Table Handler which controls the initialization, writing,  and dispersal of references for the IWageTables. This structure will also eventually handle modifications to the tables. Along with the WTH I also implemented the flyweight pattern to retrieve the references of the wage tables to be stored on the players Profession classes. This commit also includes a refactor of IProfession to Profession which touched several other files. Finally I added the pluginName gradle task to run first while running the makeall task just for fun.

+++ common/src/main/java/com/creedfreak/common/container/WageTableHandler.java
@@ -1,7 +1,6 @@
 package com.creedfreak.common.container;
 
 import com.creedfreak.common.professions.BlockTable;
-import com.creedfreak.common.professions.IProfession;
 import com.creedfreak.common.professions.IWageTable;
 import com.creedfreak.common.professions.TableType;
 import com.google.gson.reflect.TypeToken;
@@ -22,7 +21,7 @@ import java.util.logging.Logger;
 public class WageTableHandler
 {
 	public static final Type DEFAULT_WT_TYPE =
-			new TypeToken<HashMap<String, HashMap<String, Double>>> () {}.getType ();
+			new TypeToken<HashMap<String, HashMap<String, Float>>> () {}.getType ();
 
 	private static WageTableHandler Instance = new WageTableHandler ();
